refactor : component constructure
