+++ packages/assignment-7/src/components/event/calendar/EventCalendar.tsx
@@ -1,0 +1,307 @@
+import { BellIcon, ChevronLeftIcon, ChevronRightIcon } from "@chakra-ui/icons";
+import {
+  Box,
+  Heading,
+  HStack,
+  IconButton,
+  Select,
+  Table,
+  Tbody,
+  Td,
+  Text,
+  Th,
+  Thead,
+  Tr,
+  VStack,
+} from "@chakra-ui/react";
+
+import { useEffect, useState } from "react";
+import { Event, ViewType } from "../../../types/types";
+import {
+  formatMonth,
+  formatWeek,
+  getDaysInMonth,
+  getWeekDates,
+} from "../../../utils/date-utils";
+
+interface EventCalendarProps {
+  navigate: (direction: "prev" | "next") => void;
+  view: ViewType;
+  setView: React.Dispatch<React.SetStateAction<ViewType>>;
+  filteredEvents: Event[];
+  currentDate: Date;
+  notifiedEvents: number[];
+  checkUpcomingEvents: () => Promise<void>;
+}
+
+const weekDays = ["일", "월", "화", "수", "목", "금", "토"];
+
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+const fetchHolidays = (year: number, month: number) => {
+  // 실제로는 API를 호출하여 공휴일 정보를 가져와야 합니다.
+  // 여기서는 예시로 하드코딩된 데이터를 사용합니다.
+  return {
+    "2024-01-01": "신정",
+    "2024-02-09": "설날",
+    "2024-02-10": "설날",
+    "2024-02-11": "설날",
+    "2024-03-01": "삼일절",
+    "2024-05-05": "어린이날",
+    "2024-06-06": "현충일",
+    "2024-08-15": "광복절",
+    "2024-09-16": "추석",
+    "2024-09-17": "추석",
+    "2024-09-18": "추석",
+    "2024-10-03": "개천절",
+    "2024-10-09": "한글날",
+    "2024-12-25": "크리스마스",
+  };
+};
+
+const EventCalendar = ({
+  navigate,
+  view,
+  setView,
+  filteredEvents,
+  currentDate,
+  notifiedEvents,
+}: EventCalendarProps) => {
+  const [holidays, setHolidays] = useState<{ [key: string]: string }>({});
+
+  useEffect(() => {
+    const year = currentDate.getFullYear();
+    const month = currentDate.getMonth() + 1;
+    const newHolidays = fetchHolidays(year, month);
+    setHolidays(newHolidays);
+  }, [currentDate]);
+
+  return (
+    <VStack flex={1} spacing={5} align="stretch">
+      <Heading>일정 보기</Heading>
+
+      <HStack mx="auto" justifyContent="space-between">
+        <IconButton
+          aria-label="Previous"
+          icon={<ChevronLeftIcon />}
+          onClick={() => navigate("prev")}
+        />
+        <Select
+          aria-label="view"
+          value={view}
+          onChange={(e) => setView(e.target.value as "week" | "month")}
+        >
+          <option value="week">Week</option>
+          <option value="month">Month</option>
+        </Select>
+        <IconButton
+          aria-label="Next"
+          icon={<ChevronRightIcon />}
+          onClick={() => navigate("next")}
+        />
+      </HStack>
+
+      {view === "week" && (
+        <WeekCalendarView
+          currentDate={currentDate}
+          filteredEvents={filteredEvents}
+          notifiedEvents={notifiedEvents}
+        />
+      )}
+      {view === "month" && (
+        <MonthCalendarView
+          currentDate={currentDate}
+          filteredEvents={filteredEvents}
+          notifiedEvents={notifiedEvents}
+          holidays={holidays}
+        />
+      )}
+    </VStack>
+  );
+};
+
+interface CalendarProps {
+  filteredEvents: Event[];
+  currentDate: Date;
+  notifiedEvents: number[];
+  holidays: { [key: string]: string };
+}
+
+const WeekCalendarView = ({
+  currentDate,
+  filteredEvents,
+  notifiedEvents,
+}: Omit<CalendarProps, "holidays">) => {
+  return (
+    <VStack data-testid="week-view" align="stretch" w="full" spacing={4}>
+      <Heading size="md">{formatWeek(currentDate)}</Heading>
+      <Table variant="simple" w="full">
+        <Thead>
+          <Tr>
+            {weekDays.map((day) => (
+              <Th key={day} width="14.28%">
+                {day}
+              </Th>
+            ))}
+          </Tr>
+        </Thead>
+        <Tbody>
+          <Tr>
+            {getWeekDates(currentDate).map((date) => (
+              <Td
+                key={date.toISOString()}
+                height="100px"
+                verticalAlign="top"
+                width="14.28%"
+              >
+                <Text fontWeight="bold">{date.getDate()}</Text>
+                {filteredEvents
+                  .filter(
+                    (event) =>
+                      new Date(event.date).toDateString() ===
+                      date.toDateString()
+                  )
+                  .map((event) => {
+                    const isNotified = notifiedEvents.includes(event.id);
+                    return (
+                      <Box
+                        key={event.id}
+                        p={1}
+                        my={1}
+                        bg={isNotified ? "red.100" : "gray.100"}
+                        borderRadius="md"
+                        fontWeight={isNotified ? "bold" : "normal"}
+                        color={isNotified ? "red.500" : "inherit"}
+                      >
+                        <HStack spacing={1}>
+                          {isNotified && <BellIcon />}
+                          <Text fontSize="sm" noOfLines={1}>
+                            {event.title}
+                          </Text>
+                        </HStack>
+                      </Box>
+                    );
+                  })}
+              </Td>
+            ))}
+          </Tr>
+        </Tbody>
+      </Table>
+    </VStack>
+  );
+};
+
+const MonthCalendarView = ({
+  currentDate,
+  notifiedEvents,
+  filteredEvents,
+  holidays,
+}: CalendarProps) => {
+  const daysInMonth = getDaysInMonth(
+    currentDate.getFullYear(),
+    currentDate.getMonth()
+  );
+  const firstDayOfMonth = new Date(
+    currentDate.getFullYear(),
+    currentDate.getMonth(),
+    1
+  ).getDay();
+  const days = Array.from({ length: daysInMonth }, (_, i) => i + 1);
+  const weeks = [];
+  let week = Array(7).fill(null);
+
+  for (let i = 0; i < firstDayOfMonth; i++) {
+    week[i] = null;
+  }
+
+  for (const day of days) {
+    const dayIndex = (firstDayOfMonth + day - 1) % 7;
+    week[dayIndex] = day;
+    if (dayIndex === 6 || day === daysInMonth) {
+      weeks.push(week);
+      week = Array(7).fill(null);
+    }
+  }
+
+  return (
+    <VStack data-testid="month-view" align="stretch" w="full" spacing={4}>
+      <Heading size="md">{formatMonth(currentDate)}</Heading>
+      <Table variant="simple" w="full">
+        <Thead>
+          <Tr>
+            {weekDays.map((day) => (
+              <Th key={day} width="14.28%">
+                {day}
+              </Th>
+            ))}
+          </Tr>
+        </Thead>
+        <Tbody>
+          {weeks.map((week, weekIndex) => (
+            <Tr key={weekIndex}>
+              {week.map((day, dayIndex) => {
+                const dateString = day
+                  ? `${currentDate.getFullYear()}-${String(
+                      currentDate.getMonth() + 1
+                    ).padStart(2, "0")}-${String(day).padStart(2, "0")}`
+                  : "";
+                const holiday = holidays[dateString];
+
+                return (
+                  <Td
+                    key={dayIndex}
+                    height="100px"
+                    verticalAlign="top"
+                    width="14.28%"
+                    position="relative"
+                  >
+                    {day && (
+                      <>
+                        <Text fontWeight="bold">{day}</Text>
+                        {holiday && (
+                          <Text color="red.500" fontSize="sm">
+                            {holiday}
+                          </Text>
+                        )}
+                        {filteredEvents
+                          .filter(
+                            (event) => new Date(event.date).getDate() === day
+                          )
+                          .map((event) => {
+                            const isNotified = notifiedEvents.includes(
+                              event.id
+                            );
+                            return (
+                              <Box
+                                key={event.id}
+                                p={1}
+                                my={1}
+                                bg={isNotified ? "red.100" : "gray.100"}
+                                borderRadius="md"
+                                fontWeight={isNotified ? "bold" : "normal"}
+                                color={isNotified ? "red.500" : "inherit"}
+                              >
+                                <HStack spacing={1}>
+                                  {isNotified && <BellIcon />}
+                                  <Text fontSize="sm" noOfLines={1}>
+                                    {event.title}
+                                  </Text>
+                                </HStack>
+                              </Box>
+                            );
+                          })}
+                      </>
+                    )}
+                  </Td>
+                );
+              })}
+            </Tr>
+          ))}
+        </Tbody>
+      </Table>
+    </VStack>
+  );
+};
+
+export default EventCalendar;
